[INJICERT-585] re-organize packages & make obj singleton
