+++ certify-service/src/main/java/io/mosip/certify/services/ldsigner/Ed25519ProofSignature2018.java
@@ -1,10 +1,11 @@
-package io.mosip.certify.services;
+package io.mosip.certify.services.ldsigner;
 
 import com.nimbusds.jose.JWSAlgorithm;
 import info.weboftrust.ldsignatures.LdProof;
 import info.weboftrust.ldsignatures.canonicalizer.Canonicalizer;
 import info.weboftrust.ldsignatures.canonicalizer.URDNA2015Canonicalizer;
 import io.mosip.certify.core.constants.SignatureAlg;
+import io.mosip.certify.services.KeyManagerConstants;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
@@ -17,6 +18,8 @@ import java.util.Map;
 @ConditionalOnProperty(name = "mosip.certify.issuer.vc-sign-algo", havingValue = SignatureAlg.ED25519_SIGNATURE_SUITE)
 public class Ed25519ProofSignature2018 implements ProofSignatureStrategy {
 
+    Canonicalizer canonicalizer = new URDNA2015Canonicalizer();
+
     @Override
     public String getName() {
         return SignatureAlg.ED25519_SIGNATURE_SUITE;
@@ -24,7 +27,7 @@ public class Ed25519ProofSignature2018 implements ProofSignatureStrategy {
 
     @Override
     public Canonicalizer getCanonicalizer() {
-        return new URDNA2015Canonicalizer();
+        return canonicalizer;
     }
 
     @Override
