+++ certify-service/src/main/java/io/mosip/certify/services/ldsigner/RsaProofSignature2018.java
@@ -1,10 +1,11 @@
-package io.mosip.certify.services;
+package io.mosip.certify.services.ldsigner;
 
 import com.nimbusds.jose.JWSAlgorithm;
 import info.weboftrust.ldsignatures.LdProof;
 import info.weboftrust.ldsignatures.canonicalizer.Canonicalizer;
 import info.weboftrust.ldsignatures.canonicalizer.URDNA2015Canonicalizer;
 import io.mosip.certify.core.constants.SignatureAlg;
+import io.mosip.certify.services.KeyManagerConstants;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
@@ -14,6 +15,8 @@ import java.util.Map;
 @ConditionalOnProperty(name = "mosip.certify.issuer.vc-sign-algo", havingValue = SignatureAlg.RSA_SIGNATURE_SUITE)
 public class RsaProofSignature2018 implements ProofSignatureStrategy {
 
+    Canonicalizer canonicalizer = new URDNA2015Canonicalizer();
+
     @Override
     public String getName() {
         return SignatureAlg.RSA_SIGNATURE_SUITE;
@@ -21,7 +24,7 @@ public class RsaProofSignature2018 implements ProofSignatureStrategy {
 
     @Override
     public Canonicalizer getCanonicalizer() {
-        return new URDNA2015Canonicalizer();
+        return canonicalizer;
     }
 
     @Override
